API: Move conjugate_gradient_nonlinear to solvers/smooth
